hypsibius-slack/pyproject.toml
[tool.poetry]
name = "hypsibius-slack"
version = "0.1.0"
description = ""
authors = ["jonirap <joni.rapoport@gmail.com>"]
readme = "README.md"
packages = [{include = "hypsibius_slack"}]

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.21.3"
aiohttp = "^3.8.4"
environ-config = "^23.2.0"
slack-bolt = "^1.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
